Move files in LeanCode.DomainModels and it's tests to better match NS structure
