+++ src/Domain/LeanCode.DomainModels/Model/SId.cs
@@ -1,6 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
-using LeanCode.DomainModels.Serialization;
 
 namespace LeanCode.DomainModels.Model;
