+++ test/Domain/LeanCode.DomainModels.Tests/LeanCode.DomainModels.Tests.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="../../../src/Domain/LeanCode.DomainModels/LeanCode.DomainModels.csproj" />
+    <ProjectReference Include="../../../src/Domain/LeanCode.DomainModels.Generators/LeanCode.DomainModels.Generators.csproj" OutputItemType="Analyzer" />
   </ItemGroup>
 
 </Project>
