+++ test/Domain/LeanCode.DomainModels.Tests/Model/TypedIdSerializationTests.cs
@@ -2,7 +2,7 @@ using System.Text.Json;
 using LeanCode.DomainModels.Model;
 using Xunit;
 
-namespace LeanCode.DomainModels.Tests;
+namespace LeanCode.DomainModels.Tests.Model;
 
 public class TypedIdSerializationTests
 {
